// frontend/src/app/core/generated/circabc/model/dynamicPropertyDefinition.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DynamicPropertyDefinitionUpdatedValues } from './dynamicPropertyDefinitionUpdatedValues';


export interface DynamicPropertyDefinition { 
    readonly id?: string;
    readonly index?: number;
    readonly name?: string;
    /**
     * The object that is used to compile all the translations of a node into a JSON object. It is basically composed of a map with a key languaguage code and its value 
     */
    title: { [key: string]: string; };
    propertyType: DynamicPropertyDefinition.PropertyTypeEnum;
    /**
     * non mandatory field, used only when having select type 
     */
    possibleValues?: Array<string>;
    /**
     * non mandatory field, used only when having select type 
     */
    updatedValues?: Array<DynamicPropertyDefinitionUpdatedValues>;
}
export namespace DynamicPropertyDefinition {
    export const PropertyTypeEnum = {
        DateField: 'DATE_FIELD',
        TextField: 'TEXT_FIELD',
        TextArea: 'TEXT_AREA',
        Selection: 'SELECTION',
        MultiSelection: 'MULTI_SELECTION'
    } as const;
    export type PropertyTypeEnum = typeof PropertyTypeEnum[keyof typeof PropertyTypeEnum];
}